PythonGPt remembers previous chats in a session
Previously, PythonGPT here did not have a way to remember conversations, as it could only respond to one thing at a time. However, with this new code update, it can now store previous entries of a conversation and call back to them if needed. PythonGTP will not remember anything outside of the terminal session, so if you have a conversation, kill the terminal, and try to call back to certain topics, it won't remember those.